+++ pallets/loans/Cargo.toml
@@ -25,9 +25,9 @@ cfg-traits = { path = "../../libs/traits", default-features = false }
 cfg-types = { path = "../../libs/types", default-features = false }
 orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.38" }
 
-strum = { version = "0.24", default-features = false, features = ["derive"] }
 chrono = { version = "0.4", default-features = false }
 chronoutil = "0.2"
+strum = { version = "0.24", default-features = false, features = ["derive"] }
 
 # Optionals for benchmarking
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", default-features = false, optional = true, branch = "polkadot-v0.9.38" }
